+++ test/angarium/configuration_test.rb
@@ -8,7 +8,6 @@ class Angarium::ConfigurationTest < ActiveSupport::TestCase
     config = Angarium::Configuration.new
     assert_equal :default, config.job_queue
     assert_equal 10, config.http_timeout
-    assert_equal "X-Angarium-Signature", config.signature_header
     assert_equal true, config.block_private_ips
     assert_equal 5, config.retry_schedule.length
     assert_match(/Angarium/, config.user_agent)